781 pdf rubocop and linter fixes
